third_party/blink/web_tests/platform/generic/http/tests/security/subresourceIntegrity/subresource-integrity-empty-resource-blocked-expected.txt

CONSOLE ERROR: Failed to find a valid digest in the 'integrity' attribute for resource 'http://127.0.0.1:8000/security/subresourceIntegrity/resources/empty.js?block' with computed SHA-256 integrity '47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU='. The resource has been blocked.
The test passes if the zero-byte resource is blocked.
